fixes for static tenants / fallback tenants + updating readme and decision doc + relocating files
