+++ README.md
@@ -120,19 +120,22 @@ The application uses a build-time tenant generation system to provide tenant dat
 The system uses a three-tier fallback approach:
 
 1. **Fresh Database Data** (preferred) - Fetches current tenant data from the database
-2. **Seed File Data** (CI bootstrap) - Uses committed baseline data from `src/generated/static-tenants.seed.ts`
-3. **Hardcoded Fallback** (emergency) - Uses minimal fallback data from `src/scripts/fallback-tenants.ts`
+2. **Seed File Data** (CI bootstrap) - Uses committed baseline data from `src/generated/tenants/static-tenants.json`
+3. **Hardcoded Fallback** (emergency) - Uses minimal fallback data from `src/utilities/tenancy/fallback-tenants.ts`
 
 ### File Structure
 
 ```
 src/
 ├── generated/
-│   ├── static-tenants.ts      # Generated locally (gitignored)
-│   └── static-tenants.seed.ts # Committed baseline (in git)
-└── scripts/
-    ├── fallback-tenants.ts         # Hardcoded fallback data - would add new ACs over time here
-    └── generate-static-tenants.ts  # Generation script
+│   └── tenants/
+│       ├── index.ts                    # Determines which tenant (in git)
+│       └── static-tenants.json         # Committed baseline (in git)
+├── scripts/
+│   └── generate-static-tenants.ts      # Generation script
+└── utilities/
+    └── tenancy/
+        └── fallback-tenants.ts         # Hardcoded fallback data - would add new ACs over time here
 ```
 
 ### Commands
@@ -150,12 +153,12 @@ src/
 **CI/Production:**
 
 - If database is available: generates fresh data from database
-- If database unavailable: uses `static-tenants.seed.ts` (committed baseline)
+- If database unavailable: uses `static-tenants.json` (committed baseline)
 - Final fallback: uses hardcoded data from `fallback-tenants.ts`
 
 ### Updating Fallback Data
 
-When adding new tenants to production, update `src/scripts/fallback-tenants.ts` to keep the baseline current. The seed file will automatically use the updated fallback data.
+When adding new tenants to production, update `src/utilities/tenancy/fallback-tenants.ts` to keep the baseline current. The seed file will automatically use the updated fallback data.
 
 ## Developing Emails
