+++ src/utilities/tenancy/getProductionTenantUrls.ts
@@ -1,7 +1,10 @@
+import { STATIC_TENANTS } from '@/generated/tenants'
 import { getURL } from '../getURL'
-import { PRODUCTION_TENANTS, TENANTS } from './tenants'
+import { PRODUCTION_TENANTS } from './tenants'
 
 export function getProductionTenantUrls() {
-  const productionTenantDefs = TENANTS.filter((tenant) => PRODUCTION_TENANTS.includes(tenant.slug))
+  const productionTenantDefs = STATIC_TENANTS.filter((tenant) =>
+    PRODUCTION_TENANTS.includes(tenant.slug),
+  )
   return productionTenantDefs.map(({ customDomain }) => getURL(customDomain))
 }
